Modify resolvers to only resolve when a set of conditions is met
Added a _should_resolve() method to resolvers. This allows us to specify conditions for resolving, e.g. InfoboxResolver should only be used for wikipedia classes (wikipedia-class instead of wikibase-class) and PersonResolver should only resolve when the class is wikibase-person. Improved resolvers infrastructure with a decorator that invokes _should_resolve() before resolving.

NOTE: subclasses of BaseResolver should _not_ override the resolve() method. Instead, define a new method prefixed by "resolve_", e.g. "resolve_infobox". This ensures that resolvers will be wired up correctly, and the _should_resolve() function will be invoked.

Broke up StaticResolver into PersonResolver and TermResolver, and renamed ParagraphResolver to SectionsResolver.
