+++ wikipediabase/resolvers/error.py
@@ -1,6 +1,6 @@
 from wikipediabase.resolvers.base import MIN_PRIORITY, BaseResolver
 from wikipediabase.util import get_knowledgebase
-from wikipediabase.lispify import lispify
+from wikipediabase.lispify import lispify, LispType
 
 
 class ErrorResolver(BaseResolver):
@@ -17,9 +17,18 @@ class ErrorResolver(BaseResolver):
                         'kw': {'reply': repl or 'No such attribute'}},
                        typecode='error')
 
-    def resolve(self, symbol, attr, cls=None):
+    def resolve_error(self, symbol, attr, cls=None):
         kb = get_knowledgebase()
 
+        if isinstance(attr, LispType):
+            attr = attr.val
+        else:
+            attr = attr
+
+        attr = attr.lower()
+
+        # TODO: seems hacky, move to its own method and allow for multiple
+        # error conditions
         if 'wikibase-person' in kb.get_classes(symbol) and \
            attr.lower().startswith('death-') and \
            kb.get(cls, symbol, 'birth-date'):
