+++ wikipediabase/resolvers/infobox.py
@@ -1,5 +1,6 @@
+from wikipediabase.classifiers import is_wikipedia_class
 from wikipediabase.resolvers.base import BaseResolver
-from wikipediabase.util import get_infobox
+from wikipediabase.util import get_infoboxes
 from wikipediabase.lispify import lispify, LispType
 from wikipediabase.fetcher import WIKIBASE_FETCHER
 
@@ -22,7 +23,10 @@ class InfoboxResolver(BaseResolver):
         self.fetcher = kwargs.get('fetcher', WIKIBASE_FETCHER)
         self._typecode = "html"
 
-    def resolve(self, symbol, attr, cls=None):
+    def _should_resolve(self, symbol, attr, cls=None, **kwargs):
+        return is_wikipedia_class(cls)
+
+    def resolve_infobox(self, symbol, attr, cls=None):
         """
         Return the value of the attribute for the article.
         """
@@ -36,17 +40,17 @@ class InfoboxResolver(BaseResolver):
         else:
             typecode, attr = self._typecode, attr
 
-        ibox = get_infobox(symbol, self.fetcher)
+        infoboxes = get_infoboxes(symbol, cls=cls, fetcher=self.fetcher)
 
-        if ibox:
+        for ibox in infoboxes:
             result = ibox.get(attr)
             if result:
                 self.log().info("Found infobox attribute '%s'" % attr)
-                assert(isinstance(result, unicode)) # TODO: remove for production
+                assert(isinstance(result, unicode))  # TODO: remove for production
 
                 return lispify(result, typecode=typecode, infobox_attr=attr)
 
             self.log().warning("Could not find infobox attribute '%s'" % attr)
-        else:
-            self.log().warning("Could not find infobox for article '%s'"
-                               % symbol)
+
+        self.log().warning("Could not resolve attribute '%s' for '%s' with "
+                           "class '%s'", attr, symbol, cls)
